dump: make the calc fixtures deterministic and round-trippable
Three parity bugs, each hidden behind a formatting or folding step.

Data.lua:1039 sets statMap._grantedEffect inside a pairs(data.skills)
loop, and ExplosiveTrapAltX aliases ExplosiveTrap's statMap table, so
the backref was last-writer-wins over an order that varies per process
-- the lazy statMap copies then stamped whichever skill won as the mod
source, for both skills. Two dumps of the same build differed. Settle it
in sorted id order in dump_calc (matching dump_gamedata's reassign pass)
and model it in Go as GrantedEffect.StatMapOwner.

Fixtures were serialised at %.14g, which is right for compared canons
and wrong for replay input: two doubles differing in the 15th digit
print identically, and the replay then starts from a different number.
The trigger simulation turned that into a 0.1% rate difference. Add
canon.encodeExact and use it for the globalCache fixture.

Go folds untyped constant expressions at arbitrary precision, so
`1 / 0.033` is one ulp away from the reference's runtime division; that
fed a tick-rounding ceil and moved a skill duration a whole tick.

Also: processMod sets hasGlobalEffect on the granted effect it is
passed, so a calc run mutates the shared data set -- faithfully, but the
game-data canon is the post-load state, so record it at load and restore
it before comparing.

+++ data/tables.go
@@ -80,7 +80,11 @@ type Misc struct {
 func miscTable(characterConstants, monsterConstants map[string]float64) Misc {
 	return Misc{
 		ServerTickTime:                  0.033,
-		ServerTickRate:                  1 / 0.033,
+		// float64(...) forces the division to round through the double, as
+		// the reference's runtime `1 / 0.033` does: Go would otherwise fold
+		// the untyped constant expression at arbitrary precision and land
+		// one ulp away, which a tick-rounding ceil can amplify.
+		ServerTickRate:                  1 / float64(0.033),
 		AccuracyPerDexBase:              2,
 		LowPoolThreshold:                0.5,
 		TemporalChainsEffectCap:         75,
